+++ test/data-test.js
@@ -86,7 +86,8 @@ Object.keys(enemies).forEach(function (id) {
     });
   }
   if (d.drop) {
-    var VALID = ['power', 'spread', 'laser', 'heal', 'energy', 'shield', 'missile'];
+    var VALID = ['power', 'spread', 'laser', 'heal', 'energy', 'shield', 'missile',
+      'boomerang', 'option', 'life'];
     Object.keys(d.drop).forEach(function (k) {
       check(VALID.indexOf(k) >= 0, 'enemy ' + id + ' drop has unknown type ' + k);
       check(typeof d.drop[k] === 'number' && d.drop[k] >= 0 && d.drop[k] <= 1,